feat: update pkgs

+++ src/stochastic/diffusion/ou.rs
@@ -118,13 +118,13 @@ mod tests {
     plot_1d!(ou.sample(), "Fractional Ornstein-Uhlenbeck (FOU) Process");
   }
 
-  #[cfg(feature = "simd")]
+  #[cfg(all(feature = "simd", feature = "f32"))]
   #[test]
   fn sample_simd() {
     use std::time::Instant;
 
     let start = Instant::now();
-    let ou = OU::new(2.0, 1.0, 0.8, N, Some(X0), Some(1.0), None);
+    let ou = OU::new(2.0_f32, 1.0_f32, 0.8_f32, N, Some(X0 as f32), Some(1.0_f32), None);
 
     for _ in 0..100_000 {
       ou.sample_simd();
@@ -134,7 +134,7 @@ mod tests {
     println!("Elapsed time for sample_simd: {:?}", elapsed);
 
     let start = Instant::now();
-    let ou = OU::new(2.0, 1.0, 0.8, N, Some(X0), Some(1.0), None);
+    let ou = OU::new(2.0_f32, 1.0_f32, 0.8_f32, N, Some(X0 as f32), Some(1.0_f32), None);
 
     for _ in 0..100_000 {
       ou.sample();
